resource-timing/iframe-navigate-back.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing iframe navigate - back button navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    async_test(t => {
        const openee = window.open("resources/iframe-navigate-back.html");
        addEventListener("message", t.step_func_done(e => {
            assert_equals(e.data, "PASS");
        }));
        t.add_cleanup(() => openee.close());
    }, "Test that iframe navigations are not observable by the parent, even after history navigations");
</script>
